Track shared memory used by GPU process
TEST=DCHECK
BUG=79762

R=apatrick@chromium.org


Review URL: http://codereview.chromium.org/9027016

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@116366 0039d316-1c4b-4281-b951-d872f2087c98

+++ gpu/command_buffer/service/renderbuffer_manager.h
@@ -1,5 +1,5 @@
 FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
-// Copyright (c) 2011 The Chromium Authors. All rights reserved.
+// Copyright (c) 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -70,7 +70,7 @@ class RenderbufferManager {
       return has_been_bound_ && !IsDeleted();
     }
 
-    uint32 EstimatedSize();
+    size_t EstimatedSize();
 
    private:
     friend class RenderbufferManager;
@@ -159,7 +159,7 @@ class RenderbufferManager {
 
   int num_uncleared_renderbuffers_;
 
-  uint32 mem_represented_;
+  size_t mem_represented_;
 
   // Info for each renderbuffer in the system.
   typedef base::hash_map<GLuint, RenderbufferInfo::Ref> RenderbufferInfoMap;
